Added plugin structure to host plugin handle and extra data

+++ host/src/version.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <cstddef>
+#include <cstdint>
 
 namespace chimera {
 
 struct version {
-    std::size_t major = 0;
-    std::size_t minor = 0;
-    std::size_t patch = 0;
+    std::uint32_t major = 0;
+    std::uint32_t minor = 0;
+    std::uint32_t patch = 0;
 
     constexpr auto operator<=>(const version&) const -> bool = default;
 };
